refactor: moved swr into sidebar component and fixed query for load more posts
